third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_click_on_chorded_mouse_button.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Chorded button sequence L-down M-down M-up L-up
  assert_equals: received click-like events expected "auxclick,click" but got "auxclick"
[FAIL] Chorded button sequence M-down L-down L-up M-up
  assert_equals: received click-like events expected "click,auxclick" but got "click"
Harness: the test ran to completion.
